% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMF.R
\name{momf.computeRef}
\alias{momf.computeRef}
\title{Compute cell type specific mean in single cell RNA-seq data}
\usage{
momf.computeRef(sc_counts, cell_type)
}
\arguments{
\item{sc_counts}{single-cell RNA-seq data}

\item{cell_type}{the corresponding single cell labels (character)}
}
\description{
Compute cell type specific mean in single cell RNA-seq data
}
