% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMF.R
\name{momf.ct_mean}
\alias{momf.ct_mean}
\title{Compute cell type specific mean in single cell RNAseq data}
\usage{
momf.ct_mean(DataX, cell_type, ct_select = NULL, filtering = TRUE,
  min_cell = 2, min_read = 3, verbose = TRUE)
}
\description{
Compute cell type specific mean in single cell RNAseq data
}
