% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMF.R
\name{momf.fit}
\alias{momf.fit}
\title{Deconvoluation Analysis via Nonnegative Matrix Factorization}
\usage{
momf.fit(DataX, DataW = NULL, DataH = NULL, DataU = NULL,
  DataPriorU = NULL, method = "KL", rho = 2, num_iter = 5000)
}
\arguments{
\item{DataX}{nonnegative data matrix (list): bulk RNA-seq (#individuals x #genes) and scRNA-seq (#cells x #genes)}

\item{DataW}{initial nonnegative matrix (list)}

\item{DataH}{initial nonnegative matrix (list) when fitting single omic model}

\item{DataU}{initial nonnegative matrix (list) when fitting multiple omics model.}

\item{method}{two options: Kullback-Leibler (KL) divergence; Itakura-Saito (IS) divergence.
Default is "KL".}

\item{rho}{ADMM parameter, default is 2.}
}
\value{
a list.
}
\description{
The data set input \code{DataX} and \code{DataW} are necessary. \code{DataU} is needed if you do the multiple omics analysis
\code{DataH} is needed if you do the single omic analysis
Model fitting using the alternating direction method of multipliers (ADMM). 
Note: All data input should be nonnegative value, and the class should be list structure. All data sets have the same samples
}
